facebook_id migration corrected to bigInt.

+++ database/migrations/2016_04_10_004451_add_fb_id.php
@@ -13,7 +13,7 @@ class AddFbId extends Migration
     public function up()
     {
         Schema::table('users', function ($table) {
-            $table->integer('facebook_id')->unsigned();
+            $table->bigInteger('facebook_id')->unsigned();
         });
     }
